Some progress on API and offheap implementation

+++ terra-core/src/main/java/com/ritualsoftheold/terra/node/Octree.java
@@ -26,5 +26,5 @@ public interface Octree extends Node {
     
     int l_getNodeAt(int index);
     
-    void l_getData(long[] data);
+    void l_getData(int[] data);
 }
